pocketgit/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
GitPython==3.1.41
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
bcrypt==4.1.2
cryptography==42.0.5
